+++ client/src/index.jsx
@@ -3,8 +3,7 @@ import ReactDOM from 'react-dom';
 //import Waveform from './Waveform/Waveform.jsx'
 import WaveformApp from './Waveform/WaveformApp.jsx';
 // import EqualizerWindow from './EditAudio/EQ.jsx';
-import Project from './ProjectView/Project.jsx';
-import Home from './Home/Home.jsx';
+import Projects from './Projects/Projects.jsx';
 import './styles/styles.css';
 
 class App extends React.Component {
@@ -12,32 +11,18 @@ class App extends React.Component {
     super(props);
     this.state = {
       testing: false,
-      projectId: null,
     };
-
-    this.loadProject = this.loadProject.bind(this);
-  }
-
-  loadProject(event) {
-    let projectId = event.target.getAttribute('projectId');
-    this.setState({
-      projectId: projectId
-    });
   }
 
   render() {
     if (this.state.testing) {
-      if (this.state.projectId === null) {
-        return <Home loadProject={this.loadProject}/>;
-      } else {
-        return <Project projectId={this.state.projectId}/>;
-      }
+      return <Projects ownerId={'test'}/>;
     } else {
       return (
         <div>BOC
           <WaveformApp />
           <hr />
-          <button onClick={() => this.setState({testing: true})}>Load Test Project</button>
+          <button onClick={() => this.setState({testing: true})}>View Test User's Projects</button>
         </div>
       );
     }
